+++ Content.Shared/Inventory/InventorySystem.Relay.cs
@@ -61,7 +61,7 @@ public partial class InventorySystem
         SubscribeLocalEvent<InventoryComponent, BeforeEmoteEvent>(RelayInventoryEvent);
         SubscribeLocalEvent<InventoryComponent, StoodEvent>(RelayInventoryEvent);
         SubscribeLocalEvent<InventoryComponent, DownedEvent>(RelayInventoryEvent);
-        SubscribeLocalEvent<InventoryComponent, DamageChangedEvent>(RefRelayInventoryEvent);
+        SubscribeLocalEvent<InventoryComponent, DamageChangedEvent>(RefRelayInventoryEvent); // Moffstation
 
         // by-ref events
         SubscribeLocalEvent<InventoryComponent, RefreshFrictionModifiersEvent>(RefRelayInventoryEvent);
